remove legacy edge proxy and generic proxy (#9581)
