+++ localstack/aws/handlers/__init__.py
@@ -40,7 +40,6 @@ parse_service_response = service.ServiceResponseParser()
 parse_pre_signed_url_request = presigned_url.ParsePreSignedUrlRequest()
 # legacy compatibility handlers
 serve_edge_router_rules = legacy.EdgeRouterHandler()
-serve_default_listeners = legacy.DefaultListenerHandler()
 set_close_connection_header = legacy.set_close_connection_header
-pop_request_context = legacy.pop_request_context
 push_request_context = legacy.push_request_context
+pop_request_context = legacy.pop_request_context
