+++ localstack/aws/handlers/service_plugin.py
@@ -11,8 +11,6 @@ from ...utils.bootstrap import is_api_enabled
 from ..api import RequestContext
 from ..api.core import ServiceOperation
 from ..chain import Handler, HandlerChain
-from ..proxy import AwsApiListener
-from .legacy import LegacyPluginHandler
 from .service import ServiceRequestRouter
 
 LOG = logging.getLogger(__name__)
@@ -63,10 +61,7 @@ class ServiceLoader(Handler):
             if service_operation in request_router.handlers:
                 return
             if isinstance(service_plugin, Service):
-                if type(service_plugin.listener) == AwsApiListener:
-                    request_router.add_skeleton(service_plugin.listener.skeleton)
-                else:
-                    request_router.add_handler(service_operation, LegacyPluginHandler())
+                request_router.add_skeleton(service_plugin.skeleton)
             else:
                 LOG.warning(
                     f"found plugin for '{service_name}', "
