+++ localstack/services/plugins.py
@@ -10,12 +10,12 @@ from typing import Callable, Dict, List, Optional, Protocol, Tuple
 from plugin import Plugin, PluginLifecycleListener, PluginManager, PluginSpec
 
 from localstack import config
-from localstack.aws.skeleton import DispatchTable
+from localstack.aws.skeleton import DispatchTable, Skeleton
+from localstack.aws.spec import load_service
 from localstack.config import ServiceProviderConfig
 from localstack.state import StateLifecycleHook, StateVisitable, StateVisitor
 from localstack.utils.bootstrap import get_enabled_apis, is_api_enabled, log_duration
 from localstack.utils.functions import call_safe
-from localstack.utils.net import wait_for_port_status
 from localstack.utils.sync import SynchronizedDefaultDict, poll_condition
 
 # set up logger
@@ -75,16 +75,16 @@ class Service:
         self,
         name,
         start=_default,
-        check=_default,
-        listener=None,
+        check=None,
+        skeleton=None,
         active=False,
         stop=None,
         lifecycle_hook: ServiceLifecycleHook = None,
     ):
         self.plugin_name = name
         self.start_function = start
-        self.listener = listener
-        self.check_function = check if check is not _default else local_api_checker(name)
+        self.skeleton = skeleton
+        self.check_function = check
         self.default_active = active
         self.stop_function = stop
         self.lifecycle_hook = lifecycle_hook or ServiceLifecycleHook()
@@ -98,18 +98,11 @@ class Service:
             return
 
         if self.start_function is _default:
-            # fallback start method that simply adds the listener function to the list of proxy listeners if it exists
-            if not self.listener:
-                return
-
-            from localstack.services.infra import add_service_proxy_listener
-
-            add_service_proxy_listener(self.plugin_name, self.listener)
             return
 
         kwargs = {"asynchronous": asynchronous}
-        if self.listener:
-            kwargs["update_listener"] = self.listener
+        if self.skeleton:
+            kwargs["update_listener"] = self.skeleton
         return self.start_function(**kwargs)
 
     def stop(self):
@@ -160,21 +153,18 @@ class Service:
         :param service_lifecycle_hook: if left empty, the factory checks whether the provider is a ServiceLifecycleHook.
         :return: a service instance
         """
-        from localstack.aws.proxy import AwsApiListener
-
         # determine the service_lifecycle_hook
         if service_lifecycle_hook is None:
             if isinstance(provider, ServiceLifecycleHook):
                 service_lifecycle_hook = provider
 
-        # determine the delegate for injecting into the AwsApiListener
+        # determine the delegate for injecting into the skeleton
         delegate = dispatch_table_factory(provider) if dispatch_table_factory else provider
 
         service = Service(
             name=provider.service,
-            listener=AwsApiListener(provider.service, delegate=delegate),
+            skeleton=Skeleton(load_service(provider.service), delegate),
             lifecycle_hook=service_lifecycle_hook,
-            check=None,
         )
         service._provider = provider
 
@@ -702,37 +692,3 @@ def check_service_health(api, expect_shutdown=False):
         else:
             LOG.warning('Service "%s" still shutting down, retrying...', api)
         raise Exception("Service check failed for api: %s" % api)
-
-
-def local_api_checker(service: str) -> Callable:
-    """
-    Creates a health check method for the given service that works under the assumption that the real backend service
-    ports are locatable through the PROXY_LISTENER global.
-    """
-    from localstack.services.infra import PROXY_LISTENERS
-
-    if config.EAGER_SERVICE_LOADING:
-        # most services don't have a real health check, and if they would, that would dramatically increase the
-        # startup time, since health checks are done sequentially at startup. however, the health checks are needed
-        # for the lazy-loading cold start.
-        return lambda *args, **kwargs: None
-
-    def _check(expect_shutdown=False, print_error=False):
-        port = None
-        try:
-            if service not in PROXY_LISTENERS:
-                LOG.debug("cannot find backend port for service %s", service)
-                return
-            port = PROXY_LISTENERS[service][1]
-
-            if port is None:
-                # for modern ASF services, the port can be none since the service is just served by localstack
-                return
-
-            LOG.debug("checking service health %s:%d", service, port)
-            wait_for_port_status(port, expect_success=not expect_shutdown)
-        except Exception:
-            if print_error:
-                LOG.exception("service health check %s:%s failed", service, port)
-
-    return _check
