+++ localstack/testing/pytest/fixtures.py
@@ -41,7 +41,6 @@ from localstack.utils.json import CustomEncoder, json_safe
 from localstack.utils.net import wait_for_port_open
 from localstack.utils.strings import short_uid, to_str
 from localstack.utils.sync import ShortCircuitWaitException, poll_condition, retry, wait_until
-from localstack.utils.testutil import start_http_server
 
 LOG = logging.getLogger(__name__)
 
@@ -1535,13 +1534,6 @@ def acm_request_certificate(aws_client_factory):
             acm_client.delete_certificate(CertificateArn=certificate_arn)
         except Exception as e:
             LOG.debug("error cleaning up certificate %s: %s", certificate_arn, e)
-
-
-@pytest.fixture
-def tmp_http_server():
-    test_port, invocations, proxy = start_http_server()
-    yield test_port, invocations, proxy
-    proxy.stop()
 
 
 role_policy_su = {
