--- localstack/utils/server/multiserver.py
@@ -1,6 +1,0 @@
-from localstack.services import motoserver
-
-
-def get_moto_server_port():
-    # TODO: deprecated, remove
-    return motoserver.get_moto_server().port
