+++ setup.cfg
@@ -72,8 +72,6 @@ runtime =
     crontab>=0.22.6
     dnspython>=1.16.0
     docker>=6.1.1
-    flask>=3.0.0
-    flask-cors>=4.0.0
     hypercorn>=0.14.4
     json5==0.9.11
     jsonpatch>=1.24,<2.0
