+++ tests/aws/services/cloudformation/resources/test_apigateway.py
@@ -133,9 +133,8 @@ def test_cfn_apigateway_swagger_import(deploy_cfn_template, echo_http_server_pos
 
 
 @markers.aws.only_localstack
-def test_url_output(tmp_http_server, deploy_cfn_template):
-    test_port, invocations, proxy = tmp_http_server
-    integration_uri = f"http://localhost:{test_port}/{{proxy}}"
+def test_url_output(httpserver, deploy_cfn_template):
+    httpserver.expect_request("").respond_with_data(b"", 200)
     api_name = f"rest-api-{short_uid()}"
 
     stack = deploy_cfn_template(
@@ -144,7 +143,7 @@ def test_url_output(tmp_http_server, deploy_cfn_template):
         ),
         template_mapping={
             "api_name": api_name,
-            "integration_uri": integration_uri,
+            "integration_uri": httpserver.url_for("/{proxy}"),
         },
     )
