+++ tests/unit/test_misc.py
@@ -5,13 +5,11 @@ import time
 import unittest
 
 import yaml
-from requests.models import Response
 
 from localstack import config
-from localstack.services.generic_proxy import ProxyListener, start_proxy_server
 from localstack.utils import async_utils, config_listener
 from localstack.utils.aws import aws_stack
-from localstack.utils.common import TMP_FILES, download, json_safe, load_file, now_utc, parallelize
+from localstack.utils.common import json_safe, now_utc
 from localstack.utils.container_utils.container_client import PortMappings
 from localstack.utils.http import create_chunked_data, parse_chunked_data
 
@@ -148,36 +146,3 @@ class TestMisc(unittest.TestCase):
         loop.run_until_complete(asyncio.gather(*handlers))
         self.assertEqual(num_items, len(results))
         thread_pool.shutdown()
-
-
-# This test is not enabled in CI, it is just used for manual
-# testing to debug https://github.com/localstack/localstack/issues/213
-def run_parallel_download():
-    file_length = 10000000
-
-    class DownloadListener(ProxyListener):
-        def forward_request(self, method, path, data, headers):
-            sleep_time = int(path.replace("/", ""))
-            time.sleep(sleep_time)
-            response = Response()
-            response.status_code = 200
-            response._content = ("%s" % sleep_time) * file_length
-            return response
-
-    test_port = 12124
-    tmp_file_pattern = "/tmp/test.%s"
-
-    proxy = start_proxy_server(test_port, update_listener=DownloadListener())
-
-    def do_download(param):
-        tmp_file = tmp_file_pattern % param
-        TMP_FILES.append(tmp_file)
-        download("http://localhost:%s/%s" % (test_port, param), tmp_file)
-
-    values = [1, 2, 3]
-    parallelize(do_download, values)
-    proxy.stop()
-
-    for val in values:
-        tmp_file = tmp_file_pattern % val
-        assert len(load_file(tmp_file)) == file_length
